// client/src/interfaces/UserRegister.tsx
// Interface definition for UserRegister
export interface UserRegister {
    username: string;  // Property for storing username
    email: string;     // Property for storing email
    password: string;  // Property for storing password
    confirmPassword: string; // Property for confirming password
}